MOS-1443 handle double to float and small errors

+++ src/isxEventsSeries.cpp
@@ -201,15 +201,15 @@ EventsSeries::getTraceMetrics(isize_t inIndex) const
         const SpTraceMetrics_t tm = e->getTraceMetrics(inIndex);
         if (tm != nullptr)
         {
-            meanTm->m_snr += weight * tm->m_snr;
-            meanTm->m_mad += weight * tm->m_mad;
-            meanTm->m_eventRate += weight * tm->m_eventRate;
-            meanTm->m_eventAmpMedian += weight * tm->m_eventAmpMedian;
-            meanTm->m_eventAmpSD += weight * tm->m_eventAmpSD;
-            meanTm->m_riseMedian += weight * tm->m_riseMedian;
-            meanTm->m_riseSD += weight * tm->m_riseSD;
-            meanTm->m_decayMedian += weight * tm->m_decayMedian;
-            meanTm->m_decaySD += weight * tm->m_decaySD;
+            meanTm->m_snr += float(weight * tm->m_snr);
+            meanTm->m_mad += float(weight * tm->m_mad);
+            meanTm->m_eventRate += float(weight * tm->m_eventRate);
+            meanTm->m_eventAmpMedian += float(weight * tm->m_eventAmpMedian);
+            meanTm->m_eventAmpSD += float(weight * tm->m_eventAmpSD);
+            meanTm->m_riseMedian += float(weight * tm->m_riseMedian);
+            meanTm->m_riseSD += float(weight * tm->m_riseSD);
+            meanTm->m_decayMedian += float(weight * tm->m_decayMedian);
+            meanTm->m_decaySD += float(weight * tm->m_decaySD);
         }
     }
